add initial APFS support
APFS supports AFSC/decmpfs now, and the process for now appears to be identical to the HFS+ case (wacky resource forks and all); the existing code mostly works fine after adding APFS's statfs.f_type of 24 to the check for supported filesystems, so this patch does that and replaces references to "HFS+ compression" to the generic (and technically redundant (and I don't care (I am an absolute maniac))) "AFSC compression". currently I only notice a data display bug where it retrieves extremely wrong file size values when printing a verbose report on a file or directory.