More robust check for filesystem decmpfs support
Instead of checking the f_type value in statfs, now using a CFURL API to directly query whether a filesystem supports decmpfs compression. f_type values seem to be changing between OS versions (especially for APFS), so this should be a more future/past-proof approach.

Currently using a private API which has been available since 10.7, found in CFURLPriv.h in the open-source version of CoreFoundation. TODO: there's a public API for this as of macOS 10.12 (though this private one remains available and working); we should probably use dlsym to dynamically find the symbols we need for each approach, use the newer public one if it is available, and fall back to this one otherwise.