as TODO'ed: prefer public API available in Sierra
should do the same thing in either case; take a file CFURL and determine if it refers to a location on a filesystem that supports AFSC. originally this was a hardcoded check for certain values of statfs.f_type, which turned out to be unreliable particularly during the course of APFS's development when the correct vaules to check for started changing between OS releases. then I switched to a private CF API (private but non-secret, via the open source CF releases, and stable since 10.6); as of 10.12 there's an official public API for checking the same thing, so afsctool will now prefer that if it's available, because i just feel so adrift and helpless and need to feel like i'm doing things that are in some sense Right & Proper things to do even when they will almost certainly never make any real difference to anyone for any practical purposes. (TODO: make sure that sentence ends a few clauses earlier if I ever compile these into some release notes)